Split our implementation for TabDragController into separate file
There's a lot of our implementation, so it'd better to split them into
separate file
