+++ browser/ui/BUILD.gn
@@ -297,6 +297,8 @@ source_set("ui") {
       "views/tabs/brave_tab_strip.h",
       "views/tabs/brave_tab_strip_layout_helper.cc",
       "views/tabs/brave_tab_strip_layout_helper.h",
+      "views/tabs/tab_drag_controller.cc",
+      "views/tabs/tab_drag_controller.h",
       "views/tabs/vertical_tab_utils.cc",
       "views/tabs/vertical_tab_utils.h",
       "views/toolbar/bookmark_button.cc",
